pyproject.toml
[tool.poetry]
name = "first"
version = "0.1.0"
description = ""
authors = ["…"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
celery = "^5.2.7"
pyjwt = "^2.5.0"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.4"
boto3 = "^1.24.94"
pillow = "^9.2.0"
flower = "^1.2.0"
django-ses = "^3.2.0"
pika = "^1.3.1"
pydantic = "^1.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"